Atualizações na classe Controller
add controllers para renderização das páginas

+++ Projeto/bibliotecavirtual/bibliotecavirtual/src/main/java/com/senac/bibliotecavirtual/controller/EmprestimoController.java
@@ -34,7 +34,7 @@ public class EmprestimoController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Emprestimo> findById(@PathVariable Integer id) {
+    public ResponseEntity<Emprestimo> findById(@PathVariable Long id) {
         return emprestimoService.findById(id)
                 .map(emprestimo -> ResponseEntity.ok(emprestimo))
                 .orElse(ResponseEntity.notFound().build());
@@ -46,7 +46,7 @@ public class EmprestimoController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteById(@PathVariable Long id) {
         emprestimoService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
